refactor(tracing): rename `TracedRetryEventService` -> `RetryEventService`
