+++ src/main/kotlin/it/pagopa/ecommerce/eventdispatcher/services/eventretry/NotificationRetryService.kt
@@ -24,7 +24,7 @@ class NotificationRetryService(
   @Value("\${azurestorage.queues.transientQueues.ttlSeconds}")
   private val transientQueuesTTLSeconds: Int
 ) :
-  TracedRetryEventService<TransactionUserReceiptAddRetriedEvent>(
+  RetryEventService<TransactionUserReceiptAddRetriedEvent>(
     queueAsyncClient = notificationRetryQueueAsyncClient,
     retryOffset = notificationRetryOffset,
     maxAttempts = maxAttempts,
