+++ src/main/kotlin/it/pagopa/ecommerce/eventdispatcher/services/eventretry/RetryEventService.kt
@@ -7,6 +7,8 @@ import it.pagopa.ecommerce.commons.documents.v1.TransactionRetriedData
 import it.pagopa.ecommerce.commons.domain.v1.TransactionId
 import it.pagopa.ecommerce.commons.domain.v1.pojos.BaseTransaction
 import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto
+import it.pagopa.ecommerce.commons.queues.QueueEvent
+import it.pagopa.ecommerce.commons.queues.TracingInfo
 import it.pagopa.ecommerce.eventdispatcher.exceptions.NoRetryAttemptsLeftException
 import it.pagopa.ecommerce.eventdispatcher.exceptions.TooLateRetryAttemptException
 import it.pagopa.ecommerce.eventdispatcher.repositories.TransactionsEventStoreRepository
@@ -28,7 +30,11 @@ abstract class RetryEventService<E>(
   private val transientQueuesTTLSeconds: Int
 ) where E : TransactionEvent<TransactionRetriedData> {
 
-  fun enqueueRetryEvent(baseTransaction: BaseTransaction, retriedCount: Int): Mono<Void> {
+  fun enqueueRetryEvent(
+    baseTransaction: BaseTransaction,
+    retriedCount: Int,
+    tracingInfo: TracingInfo?
+  ): Mono<Void> {
     val retryEvent =
       buildRetryEvent(baseTransaction.transactionId, TransactionRetriedData(retriedCount + 1))
     val visibilityTimeout = Duration.ofSeconds((retryOffset * retryEvent.data.retryCount).toLong())
@@ -47,7 +53,7 @@ abstract class RetryEventService<E>(
             visibilityTimeout = Instant.now().plus(visibilityTimeout)))
       }
       .flatMap { storeEventAndUpdateView(it, newTransactionStatus()) }
-      .flatMap { enqueueMessage(it, visibilityTimeout) }
+      .flatMap { enqueueMessage(it, visibilityTimeout, tracingInfo) }
       .doOnError {
         logger.error(
           "Error processing retry event for transaction with id: [${retryEvent.transactionId}]", it)
@@ -75,19 +81,29 @@ abstract class RetryEventService<E>(
         viewRepository.save(it).flatMap { Mono.just(event) }
       }
 
-  private fun enqueueMessage(event: E, visibilityTimeout: Duration): Mono<Void> =
-    Mono.just(event).flatMap { eventToSend ->
-      queueAsyncClient
-        .sendMessageWithResponse(
-          BinaryData.fromObject(eventToSend),
-          visibilityTimeout,
-          Duration.ofSeconds(transientQueuesTTLSeconds.toLong()), // timeToLive
-        )
-        .doOnNext {
-          logger.info(
-            "Event: [$event] successfully sent with visibility timeout: [${it.value.timeNextVisible}] ms to queue: [${queueAsyncClient.queueName}]")
-        }
-        .then()
-        .doOnError { exception -> logger.error("Error sending event: [${event}].", exception) }
+  private fun enqueueMessage(
+    event: E,
+    visibilityTimeout: Duration,
+    tracingInfo: TracingInfo?
+  ): Mono<Void> =
+    queueAsyncClient
+      .sendMessageWithResponse(
+        queuePayload(event, tracingInfo),
+        visibilityTimeout,
+        Duration.ofSeconds(transientQueuesTTLSeconds.toLong()), // timeToLive
+      )
+      .doOnNext {
+        logger.info(
+          "Event: [$event] successfully sent with visibility timeout: [${it.value.timeNextVisible}] ms to queue: [${queueAsyncClient.queueName}]")
+      }
+      .then()
+      .doOnError { exception -> logger.error("Error sending event: [${event}].", exception) }
+
+  private fun queuePayload(event: E, tracingInfo: TracingInfo?): BinaryData {
+    return if (tracingInfo != null) {
+      BinaryData.fromObject(QueueEvent(event, tracingInfo))
+    } else {
+      BinaryData.fromObject(event)
     }
+  }
 }
